Lots of renaming and new eslint checks.
