--- src/epics/EpicsWidgets.js
@@ -1,7 +1,0 @@
-import {epicsHoc} from './EpicsHoc.js';
-import {GaugeWidget} from '../widgets/GaugeWidget.js';
-import {TextUpdateWidget} from '../widgets/TextUpdateWidget';
-
-
-export const Gauge = epicsHoc(GaugeWidget);
-export const TextUpdate = epicsHoc(TextUpdateWidget);
