+++ src/redux/epics-store.js
@@ -1,11 +1,6 @@
-//Redux API
 import {createStore, applyMiddleware} from 'redux';
-
-//Import our reducer.
-import EPICSReducer from './EPICSReducer.js';
-
-//Import our middleware from definition.
-import websocketMiddleware from './WebsocketMiddleware.js';
+import EpicsReducer from './epics-reducer';
+import websocketMiddleware from './websocket-middleware';
 
 //Create the middleware using the redux API applyMiddleware, draws from
 // the file imported above.
@@ -13,4 +8,4 @@ const middleware = applyMiddleware(websocketMiddleware);
 
 //Export the store, which is a combination of our middleware and
 //reducer.
-export const store = createStore(EPICSReducer, middleware);
+export const store = createStore(EpicsReducer, middleware);
